Update QUICKSTART.md
I found some issues with the spelling, grammar and I moved around the text to make it easier for non-contributors to follow.

In the title “Prerequistes,” there is a spelling error. It should be “Prerequisites.”
In the Unity Hub section, “Recomended” should be spelled as “Recommended.”
In the same section, “editor version” should be “editor versions,” and “you are on your own with any other version” could be clearer.
“Once Downloaded open unity hub” should have a comma: “Once downloaded, open Unity Hub.”
“navagait” should be spelled as “navigate.”
In the Installation section for Non-Contributors, “Using the dev branch might lead to Errors, Excessive Crashing, General Instability” could be rephrased for clarity.
The phrase “make note of this for when you open it in unity” should capitalize “Unity” and could be rephrased for clarity.
In the Installation section for Contributors, the instructions within the code blocks are clear, but ensure consistency in capitalization and punctuation outside of the code blocks.

I've also added a video to help the onboarding process.